Update metabolights.md
removing the bullet point “ISAcreator tool. Step by step guide for this submission method can be found in the Quick tour section of EMBL-EBI train online. This method requires the download and installation of the ISAcreator tool.” as requested by the MetaboLights Scientific Database Curator.

output of ISAcreator is no longer supported/accepted.